Improve color blob detection
